# football_academy_app/backend/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Football Academy API",
    "description": "\n    Football Academy API provides endpoints for managing player profiles, training plans, \n    challenges, achievements, and player statistics. \n    \n    This API allows players to track their progress, complete challenges, and improve their football skills.\n    ",
    "contact": {
      "name": "Football Academy Support",
      "email": "support@football-academy.app"
    },
    "license": {
      "name": "MIT License"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/challenge-progress/complete": {
      "post": {
        "tags": [
          "challenge-progress"
        ],
        "summary": "Complete a challenge",
        "description": "Record a completed challenge for the current user and award a badge if applicable",
        "operationId": "complete_challenge_challenge_progress_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeCompletionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Challenge completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeCompletion"
                },
                "example": {
                  "id": 1,
                  "player_id": 5,
                  "challenge_id": 3,
                  "completion_time": 120,
                  "score": 95,
                  "stats": {
                    "passes": 25,
                    "shots": 10,
                    "goals": 3
                  },
                  "completion_date": "2023-03-15T14:30:00"
                }
              }
            }
          },
          "404": {
            "description": "Challenge not found"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/challenge-progress/completions": {
      "get": {
        "tags": [
          "challenge-progress"
        ],
        "summary": "Get user's challenge completions",
        "description": "Retrieve all challenge completions for the currently authenticated user",
        "operationId": "get_user_completions_challenge_progress_completions_get",
        "responses": {
          "200": {
            "description": "List of all challenge completions",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChallengeCompletionWithDetails"
                  },
                  "type": "array",
                  "title": "Response Get User Completions Challenge Progress Completions Get"
                },
                "example": [
                  {
                    "id": 1,
                    "player_id": 5,
                    "challenge_id": 3,
                    "challenge": {
                      "id": 3,
                      "name": "Accuracy Challenge",
                      "description": "Complete 10 accurate passes",
                      "category": "passing"
                    },
                    "completion_time": 120,
                    "score": 95,
                    "stats": {
                      "passes": 25,
                      "shots": 10,
                      "goals": 3
                    },
                    "completion_date": "2023-03-15T14:30:00"
                  }
                ]
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/challenge-progress/completions/{challenge_id}": {
      "get": {
        "tags": [
          "challenge-progress"
        ],
        "summary": "Get completions for a specific challenge",
        "description": "Retrieve all completions for a specific challenge by the currently authenticated user",
        "operationId": "get_challenge_completions_challenge_progress_completions__challenge_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "challenge_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Challenge Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of challenge completions for the specified challenge",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChallengeCompletion"
                  },
                  "title": "Response Get Challenge Completions Challenge Progress Completions  Challenge Id  Get"
                },
                "example": [
                  {
                    "id": 1,
                    "player_id": 5,
                    "challenge_id": 3,
                    "completion_time": 120,
                    "score": 95,
                    "stats": {
                      "passes": 25,
                      "shots": 10,
                      "goals": 3
                    },
                    "completion_date": "2023-03-15T14:30:00"
                  }
                ]
              }
            }
          },
          "404": {
            "description": "No completions found for this challenge"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/challenge-progress/badges": {
      "get": {
        "tags": [
          "challenge-progress"
        ],
        "summary": "Get user's badges",
        "description": "Retrieve all badges earned by the currently authenticated user",
        "operationId": "get_user_badges_challenge_progress_badges_get",
        "responses": {
          "200": {
            "description": "List of all badges earned by the user",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BadgeWithChallenge"
                  },
                  "type": "array",
                  "title": "Response Get User Badges Challenge Progress Badges Get"
                },
                "example": [
                  {
                    "id": 1,
                    "player_id": 5,
                    "challenge_id": 3,
                    "name": "Accuracy Master",
                    "description": "Awarded for completing the Accuracy Challenge",
                    "image_url": "/badges/accuracy_badge.png",
                    "earned_at": "2023-03-15T14:30:00",
                    "challenge": {
                      "id": 3,
                      "name": "Accuracy Challenge",
                      "description": "Complete 10 accurate passes",
                      "category": "passing"
                    }
                  }
                ]
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/challenge-progress/badge-stats": {
      "get": {
        "tags": [
          "challenge-progress"
        ],
        "summary": "Get badge statistics",
        "description": "Retrieve statistics about badges earned by the currently authenticated user, grouped by category",
        "operationId": "get_badge_stats_challenge_progress_badge_stats_get",
        "responses": {
          "200": {
            "description": "Badge counts by category",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "type": "object",
                  "title": "Response Get Badge Stats Challenge Progress Badge Stats Get"
                },
                "example": {
                  "passing": 3,
                  "shooting": 2,
                  "dribbling": 1,
                  "fitness": 4
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/challenge-progress/statistics": {
      "get": {
        "tags": [
          "challenge-progress"
        ],
        "summary": "Get user's challenge statistics",
        "description": "Retrieve comprehensive statistics about challenges completed by the currently authenticated user",
        "operationId": "get_challenge_statistics_challenge_progress_statistics_get",
        "responses": {
          "200": {
            "description": "Comprehensive challenge statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Challenge Statistics Challenge Progress Statistics Get"
                },
                "example": {
                  "total_completed": 15,
                  "best_scores": {
                    "passing": {
                      "challenge_name": "Passing Master",
                      "score": 98,
                      "completion_time": 115,
                      "completion_date": "2023-03-20T16:45:00"
                    },
                    "shooting": {
                      "challenge_name": "Accuracy Shooter",
                      "score": 92,
                      "completion_time": 180,
                      "completion_date": "2023-03-18T12:30:00"
                    }
                  },
                  "recent_completions": [
                    {
                      "challenge_name": "Speed Dribbler",
                      "category": "dribbling",
                      "score": 88,
                      "completion_time": 45,
                      "completion_date": "2023-03-25T09:15:00"
                    }
                  ],
                  "category_counts": {
                    "passing": 5,
                    "shooting": 4,
                    "dribbling": 3,
                    "fitness": 3
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BadgeWithChallenge": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "badge_name": {
            "type": "string",
            "title": "Badge Name"
          },
          "challenge_id": {
            "type": "integer",
            "title": "Challenge Id"
          },
          "challenge_title": {
            "type": "string",
            "title": "Challenge Title"
          }
        },
        "type": "object",
        "required": [
          "id",
          "badge_name",
          "challenge_id",
          "challenge_title"
        ],
        "title": "BadgeWithChallenge"
      },
...
      "ChallengeCompletion": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "challenge_id": {
            "type": "integer",
            "title": "Challenge Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "challenge_id",
          "id",
          "completed_at"
        ],
        "title": "ChallengeCompletion"
      },
      "ChallengeCompletionCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "challenge_id": {
            "type": "integer",
            "title": "Challenge Id"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "challenge_id"
        ],
        "title": "ChallengeCompletionCreate"
      },
      "ChallengeCompletionWithDetails": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "challenge_id": {
            "type": "integer",
            "title": "Challenge Id"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Completed At"
          },
          "challenge_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Challenge Title"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "challenge_id",
          "completed_at",
          "challenge_title"
        ],
        "title": "ChallengeCompletionWithDetails"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  },
...
}